requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pytest
httpx
openai
python-dotenv
